Add reduction operations + javadocs

+++ src/main/java/org/deeplearning4j/examples/dataProcessing/api/analysis/columns/StringAnalysis.java
@@ -72,7 +72,7 @@ public class StringAnalysis implements ColumnAnalysis {
             return this;
         }
 
-        public Builder meanLength(int meanLength) {
+        public Builder meanLength(double meanLength) {
             this.meanLength = meanLength;
             return this;
         }
